+++ src/pulse3D/exceptions.py
@@ -59,3 +59,11 @@ class TwoPeaksInARowError(TwoFeaturesInARowError):
 
 class TooFewPeaksDetectedError(PeakDetectionError):
     pass
+
+
+class NoRecordingFilesLoadedError(Exception):
+    pass
+
+
+class SubprotocolFormatIncompatibleWithInterpolationError(Exception):
+    pass