+++ src/pulse3D/plate_recording.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-import copy
 import datetime
 import glob
 import json
@@ -26,9 +25,14 @@ from xlsxwriter.utility import xl_cell_to_rowcol
 
 from .compression_cy import compress_filtered_magnetic_data
 from .constants import *
+from .exceptions import NoRecordingFilesLoadedError
+from .exceptions import SubprotocolFormatIncompatibleWithInterpolationError
 from .magnet_finding import find_magnet_positions
 from .magnet_finding import fix_dropped_samples
 from .magnet_finding import format_well_file_data
+from .stimulation import aggregate_timepoints
+from .stimulation import create_stim_session_waveforms
+from .stimulation import realign_interpolated_stim_data
 from .transforms import apply_empty_plate_calibration
 from .transforms import apply_noise_filtering
 from .transforms import apply_sensitivity_calibration
@@ -70,12 +74,13 @@ class WellFile:
         self,
         file_path: str,
         sampling_period: Optional[Union[int, float]] = None,
-        # TODO unit test the stiffness factor, auto and override
+        # TODO unit test the stiffness factor (auto and override) and has_inverted_post_magnet?
         stiffness_factor: Optional[int] = None,
         has_inverted_post_magnet: bool = False,
     ):
         self.displacement: NDArray[(2, Any), np.float64]
         self.force: NDArray[(2, Any), np.float64]
+        self.stim_sessions: List[NDArray[(2, Any), int]] = []
 
         if stiffness_factor not in (*POST_STIFFNESS_OVERRIDE_OPTIONS, None):
             raise ValueError(
@@ -105,7 +110,7 @@ class WellFile:
                 else:
                     # calibration recordings do not have an associated barcode or post stiffness since they
                     # are creatd when a plate is not even on the istrument, so just set the stiffness factor to 1
-                    self.stiffness_factor = CALIBRATION_STIFFNESS_FACTOR  # TODO might want to make this None
+                    self.stiffness_factor = CALIBRATION_STIFFNESS_FACTOR
 
                 # extract datetimes
                 for uuid_ in (
@@ -136,16 +141,9 @@ class WellFile:
                         else None
                     )
 
-                    is_untrimmed = self.get(IS_FILE_ORIGINAL_UNTRIMMED_UUID, True)
-                    time_trimmed = None if is_untrimmed else self.attrs[TRIMMED_TIME_FROM_ORIGINAL_START_UUID]
-
                     # load sensor data. This is only possible to do here for Beta 1 data files
-                    self[TISSUE_SENSOR_READINGS] = self._load_reading(
-                        h5_file, TISSUE_SENSOR_READINGS, time_trimmed
-                    )
-                    self[REFERENCE_SENSOR_READINGS] = self._load_reading(
-                        h5_file, REFERENCE_SENSOR_READINGS, time_trimmed
-                    )
+                    self[TISSUE_SENSOR_READINGS] = self._load_reading(h5_file, TISSUE_SENSOR_READINGS)
+                    self[REFERENCE_SENSOR_READINGS] = self._load_reading(h5_file, REFERENCE_SENSOR_READINGS)
                     self._load_magnetic_data()
                 else:
                     if self.has_inverted_post_magnet:
@@ -160,6 +158,7 @@ class WellFile:
                         TIME_OFFSETS,
                         TISSUE_SENSOR_READINGS,
                         REFERENCE_SENSOR_READINGS,
+                        STIMULATION_READINGS,
                     ):
                         self[dataset] = h5_file[dataset][:]
 
@@ -188,8 +187,11 @@ class WellFile:
 
     def _load_magnetic_data(self):
         adj_raw_tissue_reading = self[TISSUE_SENSOR_READINGS].copy()
-        f = MICROSECONDS_PER_CENTIMILLISECOND if self.is_magnetic_data else MICRO_TO_BASE_CONVERSION
-        adj_raw_tissue_reading[0] *= f
+
+        time_conversion = (
+            MICROSECONDS_PER_CENTIMILLISECOND if self.is_magnetic_data else MICRO_TO_BASE_CONVERSION
+        )
+        adj_raw_tissue_reading[0] *= time_conversion
 
         # magnetic data is flipped
         if self.is_magnetic_data:
@@ -218,10 +220,11 @@ class WellFile:
             self.noise_cancelled_magnetic_data
         )
 
+        self.noise_filtered_magnetic_data: NDArray[(2, Any), int]
         if self.noise_filter_uuid is None:
-            self.noise_filtered_magnetic_data: NDArray[(2, Any), int] = self.fully_calibrated_magnetic_data
+            self.noise_filtered_magnetic_data = self.fully_calibrated_magnetic_data
         else:
-            self.noise_filtered_magnetic_data: NDArray[(2, Any), int] = apply_noise_filtering(
+            self.noise_filtered_magnetic_data = apply_noise_filtering(
                 self.fully_calibrated_magnetic_data, self.filter_coefficients
             )
 
@@ -246,7 +249,7 @@ class WellFile:
             self.displacement, stiffness_factor=self.stiffness_factor, in_mm=False
         )
 
-    def get(self, key, default):
+    def get(self, key, default=None):
         try:
             return self[key]
         except Exception:
@@ -301,55 +304,30 @@ class WellFile:
             tzinfo=datetime.timezone.utc
         )
 
-    def _load_reading(self, h5_file, reading_type: str, time_trimmed) -> NDArray[(Any, Any), int]:
-        recording_start_index = self[START_RECORDING_TIME_INDEX_UUID]
-        beginning_data_acquisition_ts = self[UTC_BEGINNING_DATA_ACQUISTION_UUID]
-
-        if reading_type == REFERENCE_SENSOR_READINGS:
-            initial_timestamp = self[UTC_FIRST_REF_DATA_POINT_UUID]
-            sampling_period = self[REF_SAMPLING_PERIOD_UUID]
-        else:
-            initial_timestamp = self[UTC_FIRST_TISSUE_DATA_POINT_UUID]
-            sampling_period = self[TISSUE_SAMPLING_PERIOD_UUID]
-
-        recording_start_index_useconds = int(recording_start_index) * MICROSECONDS_PER_CENTIMILLISECOND
-        timestamp_of_start_index = beginning_data_acquisition_ts + datetime.timedelta(
-            microseconds=recording_start_index_useconds
-        )
-
-        time_delta = initial_timestamp - timestamp_of_start_index
-        time_delta_centimilliseconds = int(
-            time_delta / datetime.timedelta(microseconds=MICROSECONDS_PER_CENTIMILLISECOND)
-        )
-
+    def _load_reading(self, h5_file, reading_type: str) -> NDArray[(Any, Any), int]:
+        sampling_period = self[
+            REF_SAMPLING_PERIOD_UUID
+            if reading_type == REFERENCE_SENSOR_READINGS
+            else TISSUE_SAMPLING_PERIOD_UUID
+        ]
         time_step = int(sampling_period / MICROSECONDS_PER_CENTIMILLISECOND)
 
-        # adding `[:]` loads the data as a numpy array giving us more flexibility of multi-dimensional arrays
-        data = h5_file[reading_type][:]
-        if len(data.shape) == 1:
-            data = data.reshape(1, data.shape[0])
+        tissue_contraction_amplitudes = h5_file[reading_type][:]
+        num_data_points = tissue_contraction_amplitudes.shape[-1]
 
-        # fmt: off
-        # black reformatted this into a few very ugly lines of code
-        times = np.mgrid[: data.shape[1],] * time_step
-        # fmt: on
+        timepoints = np.arange(num_data_points) * time_step
 
-        if time_trimmed:
-            new_times = times + time_delta_centimilliseconds
-            start_index = _find_start_index(time_trimmed, new_times)
-            time_delta_centimilliseconds = int(new_times[start_index])
-
-        return np.concatenate((times + time_delta_centimilliseconds, data), dtype=np.int32)
+        return np.array([timepoints, tissue_contraction_amplitudes], dtype=np.int32)
 
 
 class PlateRecording:
     def __init__(
         self,
         path,
-        force_df: pd.DataFrame = None,
+        recording_df: pd.DataFrame = None,
         start_time: Union[float, int] = 0,
         end_time: Optional[Union[float, int]] = None,
-        # TODO unit test the stiffness factor, auto and override
+        # TODO unit test the stiffness factor (auto and override), inverted_post_magnet_wells
         stiffness_factor: Optional[int] = None,
         inverted_post_magnet_wells: Optional[List[str]] = None,
     ):
@@ -368,6 +346,8 @@ class PlateRecording:
         self.start_time_secs = start_time
         self.end_time_secs = end_time
 
+        self._created_from_dataframe = recording_df is not None
+
         if self.path.endswith(".zip"):
             with tempfile.TemporaryDirectory() as tmpdir:
                 zf = zipfile.ZipFile(path)
@@ -385,17 +365,21 @@ class PlateRecording:
                 self.path, stiffness_factor, inverted_post_magnet_wells
             )
 
-        if not self.wells:
-            # might not be any well files in the path
-            # TODO should probably raise an error here
-            return
+        # make sure at least one WellFile was loaded
+        if not any(self.wells):
+            raise NoRecordingFilesLoadedError()
 
         # currently file versions 1.0.0 and above must have all their data processed together
         if not self.is_optical_recording and self.wells[0].version >= VersionInfo.parse("1.0.0"):
-            if force_df is None:
-                self._process_plate_data(calibration_recordings)
+            if self._created_from_dataframe:
+                self._load_dataframe(recording_df)
             else:
-                self._load_dataframe(force_df)
+                self._process_plate_data(calibration_recordings)
+
+                if self.wells[0][FILE_FORMAT_VERSION_METADATA_KEY] >= VersionInfo.parse(
+                    MIN_FILE_VERSION_FOR_STIM_INTERPOLATION
+                ):
+                    self._process_stim_data()
 
     def _process_plate_data(self, calibration_recordings):
         if not all(isinstance(well_file, WellFile) for well_file in self.wells) or len(self.wells) != 24:
@@ -438,8 +422,7 @@ class PlateRecording:
 
         # create displacement and force arrays for each WellFile
         log.info("Create diplacement and force data for each well")
-        for well_idx in range(24):
-            well_file = self.wells[well_idx]
+        for well_idx, well_file in enumerate(self):
             x = estimated_magnet_positions["X"][:, well_idx]
             if flip_data:
                 x *= -1
@@ -453,18 +436,68 @@ class PlateRecording:
                 well_file.displacement, stiffness_factor=well_file.stiffness_factor
             )
 
+    def _process_stim_data(self) -> None:
+        log.info("Interpolating stim sessions")
+
+        start_time_us = int(self.start_time_secs * MICRO_TO_BASE_CONVERSION)
+        end_time_us = (
+            int(self.end_time_secs * MICRO_TO_BASE_CONVERSION)
+            if self.end_time_secs
+            else self.wells[0][TIME_INDICES][-1]
+        )
+
+        for wf in self:
+            if not wf[STIMULATION_READINGS].shape[-1]:
+                continue
+
+            stim_protocol = json.loads(wf[STIMULATION_PROTOCOL_UUID])
+
+            try:
+                stim_sessions_waveforms = create_stim_session_waveforms(
+                    stim_protocol["subprotocols"], wf[STIMULATION_READINGS], start_time_us, end_time_us
+                )
+            except SubprotocolFormatIncompatibleWithInterpolationError:
+                log.info("Subprotocol format not supported by intperpolation")
+                return
+
+            charge_conversion_factor = (
+                MILLI_TO_BASE_CONVERSION if stim_protocol["stimulation_type"] == "C" else 1
+            )
+
+            for waveform in stim_sessions_waveforms:
+                if not waveform.shape[-1]:
+                    continue
+                waveform[0] -= wf[TIME_INDICES][0]
+                waveform[1] //= charge_conversion_factor
+                wf.stim_sessions.append(waveform)
+
     def _load_dataframe(self, df: pd.DataFrame) -> None:
-        """Add time and force data to well files in PR.
+        """Add time and force data to well files in PlateRecording.
 
         Args:
             df: pd.Dataframe existing time force data to be added
         """
-        time = df["Time (s)"].values
+        force_timepoints = df["Time (s)"].values
 
-        for well in self.wells:
-            well_name = well[WELL_NAME_UUID]
-            well_data = np.vstack((time, df[f"{well_name}__raw"])).astype(np.float64)
-            well.force = well_data
+        try:
+            stim_timepoints = df["Stim Time (µs)"].values
+        except KeyError:
+            stim_timepoints = None
+
+        for wf in self.wells:
+            well_name = wf[WELL_NAME_UUID]
+            raw_force_amplitudes = df[f"{well_name}__raw"]
+            wf.force = np.vstack((force_timepoints, raw_force_amplitudes)).astype(np.float64)
+
+            if stim_timepoints is None:
+                continue
+
+            stim_col_titles = sorted([col for col in df if f"{well_name}__stim" in col])
+
+            for col_title in stim_col_titles:
+                stim_session_raw = np.array([stim_timepoints, df[col_title]])
+                stim_session = stim_session_raw[:, ~np.isnan(stim_session_raw[1])].astype(int)
+                wf.stim_sessions.append(stim_session)
 
     def write_time_force_csv(self, output_dir: str):
         # get recording name
@@ -478,16 +511,16 @@ class PlateRecording:
 
         force_data = dict({"Time (s)": pd.Series(truncated_time_sec)})
 
-        for well in self.wells:
+        for well in self:
             well_name = well.get(WELL_NAME_UUID, None)
             force_data[well_name] = pd.Series(well.force[1])
 
-        time_force_df = pd.DataFrame(force_data)
-        time_force_df.to_csv(output_path, index=False)
+        time_recording_df = pd.DataFrame(force_data)
+        time_recording_df.to_csv(output_path, index=False)
 
-        return time_force_df, output_path
+        return time_recording_df, output_path
 
-    def to_dataframe(self, from_dataframe: bool = False) -> pd.DataFrame:
+    def to_dataframe(self) -> pd.DataFrame:
         """Creates DataFrame from PlateRecording with all the data
         interpolated, normalized, and scaled. The returned dataframe contains
         one column for time in ms and one column for each well.
@@ -500,55 +533,77 @@ class PlateRecording:
         >>> peak_detector(df[['time', 'A1']].transpose().values)
 
         The dataframe needs to be transposed before converting to NDArray because peak_detector
-        wants an ndarray of shape (2,N) while dataframe[['time', 'A1']] is in shape (N,2)
+        wants an ndarray of shape (2, N) while dataframe[['time', 'A1']] is in shape (N,2)
         """
-        data = {}
+        # get first valid well and set interpolation period. Creating new iter to be safe
+        first_well = next(iter(self))
 
-        # get first valid well and set interpolation period
-        first_well = [pw for pw in self.wells if pw][0]
+        # add interpolated force timepoints
+        if self._created_from_dataframe:
+            raise NotImplementedError("Cannot export a DF if created from a DF. Just use the original")
 
-        if self.is_optical_recording:
-            interp_period = first_well[INTERPOLATION_VALUE_UUID]
-        else:
-            interp_period = INTERPOLATED_DATA_PERIOD_US
+        min_time = min([wf.force[0][0] for wf in self])
+        max_time = max([wf.force[0][-1] for wf in self])
+        interp_period = (
+            first_well[INTERPOLATION_VALUE_UUID] if self.is_optical_recording else INTERPOLATED_DATA_PERIOD_US
+        )
+        interp_timepoints = np.arange(min_time, max_time + interp_period, interp_period)
 
-        if from_dataframe:
-            time_steps = copy.deepcopy(first_well.force[0])
-        else:
-            max_time = max([w.force[0][-1] for w in self.wells if w])
-            min_time = min([w.force[0][0] for w in self.wells if w])
-            time_steps = np.arange(min_time, max_time + interp_period, interp_period)
+        data = {"Time (s)": pd.Series(interp_timepoints)}
 
-        data["Time (s)"] = pd.Series(time_steps)
+        output_stim_data = first_well.version >= VersionInfo.parse("1.0.0")
 
-        for i, w in enumerate(self.wells):
-            if not from_dataframe:
-                start_idx, end_idx = truncate(
-                    source_series=time_steps, lower_bound=w.force[0][0], upper_bound=w.force[0][-1]
-                )
+        # add stim timepoints
+        if output_stim_data:
+            aggregate_stim_timepoints_us = aggregate_timepoints(
+                [session_data[0] for wf in self for session_data in wf.stim_sessions]
+            )
+            aggregate_stim_timepoints_us_for_plotting = np.repeat(aggregate_stim_timepoints_us, 2)
+            data["Stim Time (µs)"] = pd.Series(aggregate_stim_timepoints_us_for_plotting)
 
-                interp_fn = interpolate.interp1d(w.force[0, :], w.force[1, :])
-                interp_force = interp_fn(time_steps[start_idx : end_idx + 1])
+        # iterating over self.wells instead of using __iter__ so well_idx is preserved
+        for well_idx, wf in enumerate(self.wells):
+            if not wf:
+                continue
 
-            else:
-                interp_force = copy.deepcopy(w.force[1, 1:])
+            well_name = wf.get(WELL_NAME_UUID, TWENTY_FOUR_WELL_PLATE.get_well_name_from_well_index(well_idx))
 
-            data[f"{w.get(WELL_NAME_UUID, str(i))}__raw"] = pd.Series(w.force[1, :])
+            # add raw force data
+            data[f"{well_name}__raw"] = pd.Series(wf.force[1, :])
 
-            min_value = min(interp_force)
-            interp_force -= min_value
-            interp_force *= MICRO_TO_BASE_CONVERSION
+            # add unit adjusted + normalized force data
+            start_idx, end_idx = truncate(
+                source_series=interp_timepoints, lower_bound=wf.force[0, 0], upper_bound=wf.force[0, -1]
+            )
 
-            data[w.get(WELL_NAME_UUID, str(i))] = pd.Series(interp_force)
+            interp_fn = interpolate.interp1d(wf.force[0, :], wf.force[1, :])
+            interp_force_unewtons = interp_fn(interp_timepoints[start_idx : end_idx + 1])
 
-        return pd.DataFrame(data).dropna()
+            min_value = min(interp_force_unewtons)
+
+            interp_force_newtons_normalized = (interp_force_unewtons - min_value) * MICRO_TO_BASE_CONVERSION
+            data[well_name] = pd.Series(interp_force_newtons_normalized)
+
+            # add stim data
+            if output_stim_data:
+                for i, session_data in enumerate(wf.stim_sessions):
+                    data[f"{well_name}__stim_{i}"] = pd.Series(
+                        realign_interpolated_stim_data(
+                            aggregate_stim_timepoints_us_for_plotting, session_data
+                        )
+                    )
+
+        df = pd.DataFrame(data)
+        if not output_stim_data:
+            df.dropna(inplace=True)
+
+        return df
 
     @staticmethod
     def from_dataframe(path, df: pd.DataFrame):
-        # multi zip files
         # only allowed for one recording at a time assuming a user would only ever pass a dataframe to one recording
         log.info(f"Loading recording from file {os.path.basename(path)}")
-        yield PlateRecording(path, force_df=df)
+        yield PlateRecording(path, recording_df=df)
 
     @staticmethod
     def from_directory(path, **kwargs):